change endpoints to hyphens url structure
