+++ src/verify-token.js
